+++ src/main/java/com/rbkmoney/threeds/server/ds/client/impl/RBKMoneyPlatformDsClient.java
@@ -2,8 +2,11 @@ package com.rbkmoney.threeds.server.ds.client.impl;
 
 import com.rbkmoney.threeds.server.config.properties.EnvironmentProperties;
 import com.rbkmoney.threeds.server.converter.MessageToErrorResConverter;
+import com.rbkmoney.threeds.server.domain.root.Message;
 import com.rbkmoney.threeds.server.flow.ErrorCodeResolver;
 import com.rbkmoney.threeds.server.flow.ErrorMessageResolver;
+import org.springframework.web.client.ResourceAccessException;
+import org.springframework.web.client.RestClientException;
 import org.springframework.web.client.RestTemplate;
 
 public class RBKMoneyPlatformDsClient extends AbstractDsClient {
@@ -15,5 +18,16 @@ public class RBKMoneyPlatformDsClient extends AbstractDsClient {
             ErrorCodeResolver errorCodeResolver,
             ErrorMessageResolver errorMessageResolver) {
         super(restTemplate, environmentProperties, messageToErrorConverter, errorCodeResolver, errorMessageResolver);
+    }
+
+    @Override
+    public Message request(Message requestMessage) {
+        try {
+            return processHttpPost(requestMessage).getBody();
+        } catch (ResourceAccessException ex) {
+            return getMessage(ex, requestMessage, errorCodeResolver.resolve(ex));
+        } catch (RestClientException ex) {
+            return getMessage(ex, requestMessage, errorCodeResolver.resolve(ex));
+        }
     }
 }
